Fix : 알람 API TopNav에서 관리
